+++ .gcloudignore
@@ -1,0 +1,9 @@
+.gcloudignore
+.git
+.gitignore
+
+.idea/
+**/__pycache__/
+
+db.json
+*.md
