+++ app.yaml
@@ -1,5 +1,5 @@
 runtime: python39
-entrypoint: gunicorn -b :8080 -w 2 tgbot:app
+entrypoint: gunicorn -b :8080 -w 2 tgbot.bot:app
 
 service: codeforcewarrior-bot
