--- cf_update/requirements.txt
@@ -1,11 +1,0 @@
-requests
-pydantic
-cachetools
-aiohttp[speedups]
-aiocache
-git+https://github.com/jazzband/prettytable#egg=prettytable
-aiocron
-aiotinydb
-tinydb==3.12.1
-ujson
-python-telegram-bot==20.0a2
