+++ main.py
@@ -1,0 +1,1 @@
+from tgbot.functions import cf_verification, decline_join_request
