+++ requirements.txt
@@ -8,6 +8,7 @@ google-cloud-firestore
 functions-framework==3.*
 Flask==2.0.2
 
+# TG bot + functions
 gunicorn
 requests
 pydantic
@@ -15,3 +16,10 @@ cachetools
 aiohttp[speedups]
 aiocache
 git+https://github.com/jazzband/prettytable#egg=prettytable
+
+# cf_update
+aiocron
+aiotinydb
+tinydb==3.12.1
+ujson
+python-telegram-bot==20.0a2
