+++ tgbot/cf_update/stickers.py
