+++ tgbot/clist/client.py
@@ -3,8 +3,8 @@ from datetime import timedelta
 
 import requests
 
-from .models import ContestInfo
-from .utils import ClistError, RESOURCES, hkt_now
+from tgbot.clist.models import ClistError, ContestInfo
+from tgbot.utils import RESOURCES, hkt_now
 
 logger = logging.getLogger(__name__)
