+++ tgbot/codeforces/__init__.py
@@ -1,0 +1,3 @@
+from tgbot.codeforces.async_client import AsyncCodeforcesAPI
+from tgbot.codeforces.client import CodeforcesAPI
+from tgbot.codeforces.models import *
